+++ src/assets/changelog.md
@@ -3,6 +3,7 @@
 - Improve inputs to have labels instead of placeholders.
 - Improve player shuffle logic and design.
 - Add song to selected playlist.
+- Add link for categories (share link of categories)
 - Fix issue when adding a new song.
 - Fix player popup is visible when player is expanded.